adding the basics of the weather component, needs more work.
