+++ .env
@@ -1,0 +1,2 @@
+VITE_APP_ID:"95256c102297fc10c26b6bf4ed2e636a"
+"
