+++ src/components/Weather.css
@@ -1,0 +1,77 @@
+.weather {
+    place-self: center;
+    padding: 40px;
+    border-radius: 10px;
+    background-image: linear-gradient(45deg, #2f4680, #500ae4);
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 650px;
+    /* i want to move the search bar to the top right of the page */
+}
+.search-bar {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+}
+.search-bar input {
+    height: 50px;
+    border: none;
+    outline: none;
+    border-radius: 40px;
+    padding-left: 25px;
+    color: #626262;
+    background: #ebfffc;
+    font-size: 18px;
+}
+
+.search-bar img{
+    width: 15px;
+    padding: 15px;
+    border-radius: 50%;
+    background: #ebfffc;
+    cursor: pointer;
+}
+
+.weather-icon {
+    width: 150px;
+    margin: 30px 0;
+}
+
+.temperature{
+    color: white;
+    font-size: 50px;
+    line-height: .1;
+}
+
+.location{
+    color: white;
+    font-size: 50px;
+    line-height: .1;
+}
+.weather-data {
+    width: 100%;
+    margin-top: 40px;
+    color: #fff;
+    display: flex;
+    justify-content: space-between;
+}
+
+.weather-data .col {
+    display: flex;
+    align-items: flex-start;
+    gap: 12px;
+    font-size: 22px;
+}
+.weather-data .col span {
+    display: block;
+    font-size: 16px;
+}
+.weather-data .col span {
+    display: block;
+    font-size: 16px;
+}
+.weather-data .col img {
+    width: 26px;
+    margin-top: 10px;
+}
