+++ src/index.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import Weather from './components/Weather';
 import './index.css';
-import DominantColor from './DominantColor';
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <DominantColor />
-  </React.StrictMode>
+  <div>
+    <Weather />
+  </div>
 );
 
 // If you want to start measuring performance in your app, pass a function
